Fixed windows/clang build

+++ src/types/Buffer.cpp
@@ -178,12 +178,12 @@ const std::string Buffer::ReadString() {
 	}
 }
 
-void Buffer::WriteVec2u( const Vec2< long unsigned int > val ) {
+void Buffer::WriteVec2u( const Vec2< unsigned long long > val ) {
 	WriteImpl( T_VEC2U, (const char*)&val, sizeof( val ) );
 }
 
-const Vec2< long unsigned int > Buffer::ReadVec2u() {
-	Vec2< long unsigned int > val = { 0, 0 };
+const Vec2< unsigned long long > Buffer::ReadVec2u() {
+	Vec2< unsigned long long > val = { 0, 0 };
 	size_t sz = 0;
 	ReadImpl( T_VEC2U, (char*)&val, &sz, sizeof( val ) );
 	return val;
